+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.8.0] - Unreleased
+
+### Added
+
+- Added **rich_measure** for tree
+- Added rich.align.VerticalCenter
+- Added height to ConsoleOptions
+
 ## [9.7.0] - 2021-01-09
 
 ### Added